+++ app/tests/Invoice/InvoiceServicePaymentMethodTest.php
@@ -18,6 +18,7 @@ use App\Enum\Channel;
 use App\Enum\ReservationStatus;
 use App\Invoice\InvoiceNumber;
 use App\Invoice\InvoiceNumberAllocator;
+use App\Invoice\InvoiceNumberFormat;
 use App\Invoice\InvoicePdfRenderer;
 use App\Invoice\InvoiceService;
 use App\Invoice\IssuerProfileProvider;
@@ -42,6 +43,8 @@ final class InvoiceServicePaymentMethodTest extends TestCase
 
         $allocator = $this->createMock(InvoiceNumberAllocator::class);
         $allocator->method('allocate')->willReturn(new InvoiceNumber(2026, 12));
+
+        $numberFormat = new InvoiceNumberFormat($this->createMock(SettingRepository::class));
 
         $pdfRenderer = $this->createMock(InvoicePdfRenderer::class);
         $pdfRenderer->method('renderToFile')->willReturn('/tmp/test-invoice.pdf');
@@ -71,6 +74,7 @@ final class InvoiceServicePaymentMethodTest extends TestCase
             $em,
             $invoiceRepo,
             $allocator,
+            $numberFormat,
             $pdfRenderer,
             $spayd,
             $this->createMock(CnbExchangeRateClient::class),